Revert "Merge branch 'master' of https://github.com/cmangos/mangos-classic"
This reverts commit 5d8632d906cb0cc358d659fb7e3fa46a5248aa35, reversing
changes made to a417afed9e6a62f54c2a470c9eb0490d6b9b4fbc.

+++ dep/src/g3dlite/System.cpp
@@ -226,7 +226,7 @@ void System::init() {
         }
 
         m_numCores = systemInfo.dwNumberOfProcessors;
-        uintptr_t maxAddr = (uintptr_t)systemInfo.lpMaximumApplicationAddress;
+        uint32 maxAddr = (uint32)systemInfo.lpMaximumApplicationAddress;
         {
             char c[1024];
             sprintf(c, "%d x %d-bit %s processor",
